[WFLY-18027] moved package org.wildfly.quickstarts.todos to org.jboss.as.quickstarts.todos
