+++ todo-backend/src/main/java/org/jboss/as/quickstarts/todos/ToDoController.java
@@ -19,7 +19,7 @@
  * Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
-package org.wildfly.quickstarts.todos;
+package org.jboss.as.quickstarts.todos;
 
 import static jakarta.ws.rs.core.MediaType.APPLICATION_JSON;
